+++ amethyst_assets/src/lib.rs
@@ -22,8 +22,11 @@ extern crate log;
 extern crate parking_lot;
 extern crate rayon;
 extern crate ron;
+#[macro_use]
 extern crate serde;
 extern crate shred;
+#[macro_use]
+extern crate shred_derive;
 
 #[macro_use]
 #[cfg(feature = "profiler")]
@@ -34,6 +37,7 @@ pub use cache::Cache;
 pub use error::{Error, ErrorKind, Result, ResultExt};
 pub use formats::RonFormat;
 pub use loader::Loader;
+pub use prefab::{AssetPrefab, Prefab, PrefabData, PrefabError, PrefabLoader, PrefabLoaderSystem};
 pub use progress::{Completion, Progress, ProgressCounter, Tracker};
 pub use reload::{HotReloadBundle, HotReloadStrategy, HotReloadSystem, Reload, SingleFile};
 pub use source::{Directory, Source};
@@ -44,6 +48,7 @@ mod cache;
 mod error;
 mod formats;
 mod loader;
+mod prefab;
 mod progress;
 mod reload;
 mod source;