+++ amethyst_controls/src/lib.rs
@@ -3,6 +3,8 @@ extern crate amethyst_assets;
 extern crate amethyst_core;
 extern crate amethyst_input;
 extern crate amethyst_renderer;
+#[macro_use]
+extern crate serde;
 extern crate winit;
 
 #[cfg(feature = "profiler")]
@@ -14,7 +16,7 @@ mod resources;
 mod systems;
 
 pub use self::bundles::{ArcBallControlBundle, FlyControlBundle};
-pub use self::components::{ArcBallControlTag, FlyControlTag};
+pub use self::components::{ArcBallControlTag, ControlTagPrefab, FlyControlTag};
 pub use self::resources::WindowFocus;
 pub use self::systems::{ArcBallMovementSystem, CursorHideSystem, FlyMovementSystem,
                         FreeRotationSystem, MouseFocusUpdateSystem};