+++ amethyst_utils/src/lib.rs
@@ -1,5 +1,7 @@
 extern crate amethyst_assets;
+extern crate amethyst_controls;
 extern crate amethyst_core;
+extern crate amethyst_renderer;
 #[macro_use]
 extern crate serde;
 extern crate shred;
@@ -11,4 +13,5 @@ extern crate thread_profiler;
 
 pub mod circular_buffer;
 pub mod fps_counter;
+pub mod scene;
 pub mod tag;