+++ examples/arc_ball_camera/main.rs
@@ -2,48 +2,24 @@
 
 extern crate amethyst;
 
-use amethyst::assets::Loader;
-use amethyst::controls::{ArcBallControlBundle, ArcBallControlTag, FlyControlTag};
-use amethyst::core::cgmath::{Deg, Vector3};
-use amethyst::core::transform::{GlobalTransform, Transform, TransformBundle};
-use amethyst::ecs::prelude::{Entity, World};
+use amethyst::assets::{PrefabLoader, PrefabLoaderSystem, RonFormat};
+use amethyst::controls::ArcBallControlBundle;
+use amethyst::core::transform::TransformBundle;
 use amethyst::input::{is_close_requested, is_key, InputBundle};
-use amethyst::renderer::{AmbientColor, Camera, DrawShaded, Event, Material, MaterialDefaults,
-                         MeshHandle, ObjFormat, PosNormTex, Projection, Rgba, VirtualKeyCode};
+use amethyst::renderer::{DrawShaded, Event, PosNormTex, VirtualKeyCode};
+use amethyst::utils::scene::BasicScenePrefab;
 use amethyst::{Application, Error, GameData, GameDataBuilder, State, StateData, Trans};
+
+type MyPrefabData = BasicScenePrefab<Vec<PosNormTex>>;
 
 struct ExampleState;
 
 impl<'a, 'b> State<GameData<'a, 'b>> for ExampleState {
     fn on_start(&mut self, data: StateData<GameData>) {
-        let StateData { world, .. } = data;
-        let assets = load_assets(&world);
-
-        // Add cube to scene
-        let mut trans = Transform::default();
-        trans.translation = Vector3::new(-5.0, 0.0, -10.0);
-        world
-            .create_entity()
-            .with(assets.cube.clone())
-            .with(assets.red.clone())
-            .with(trans)
-            .with(GlobalTransform::default())
-            .build();
-
-        let mut trans = Transform::default();
-        trans.translation = Vector3::new(0.0, 0.0, -5.0);
-
-        let entity = world
-            .create_entity()
-            .with(assets.cube.clone())
-            .with(assets.red.clone())
-            .with(trans)
-            .with(GlobalTransform::default())
-            .build();
-
-        initialise_camera(world, entity);
-
-        world.add_resource(AmbientColor(Rgba::from([0.1; 3])));
+        let prefab_handle = data.world.exec(|loader: PrefabLoader<MyPrefabData>| {
+            loader.load("prefab/arc_ball_camera.ron", RonFormat, (), ())
+        });
+        data.world.create_entity().with(prefab_handle).build();
     }
 
     fn handle_event(&mut self, _: StateData<GameData>, event: Event) -> Trans<GameData<'a, 'b>> {
@@ -60,28 +36,6 @@ impl<'a, 'b> State<GameData<'a, 'b>> for ExampleState {
     }
 }
 
-struct Assets {
-    cube: MeshHandle,
-    red: Material,
-}
-
-fn load_assets(world: &World) -> Assets {
-    let mesh_storage = world.read_resource();
-    let tex_storage = world.read_resource();
-    let mat_defaults = world.read_resource::<MaterialDefaults>();
-    let loader = world.read_resource::<Loader>();
-
-    let red = loader.load_from_data([1.0, 0.0, 0.0, 1.0].into(), (), &tex_storage);
-    let red = Material {
-        albedo: red,
-        ..mat_defaults.0.clone()
-    };
-
-    let cube = loader.load("mesh/cube.obj", ObjFormat, (), (), &mesh_storage);
-
-    Assets { cube, red }
-}
-
 fn main() -> Result<(), Error> {
     let resources_directory = format!("{}/examples/assets", env!("CARGO_MANIFEST_DIR"));
 
@@ -96,6 +50,7 @@ fn main() -> Result<(), Error> {
     );
 
     let game_data = GameDataBuilder::default()
+        .with(PrefabLoaderSystem::<MyPrefabData>::default(), "", &[])
         .with_bundle(TransformBundle::new().with_dep(&[]))?
         .with_bundle(
             InputBundle::<String, String>::new().with_bindings_from_file(&key_bindings_path),
@@ -105,22 +60,4 @@ fn main() -> Result<(), Error> {
     let mut game = Application::build(resources_directory, ExampleState)?.build(game_data)?;
     game.run();
     Ok(())
-}
-
-fn initialise_camera(world: &mut World, entity: Entity) {
-    let local = Transform::default();
-
-    world
-        .create_entity()
-        .with(Camera::from(Projection::perspective(1.3, Deg(60.0))))
-        .with(local)
-        .with(GlobalTransform::default())
-        // Will allow the FreeRotationSystem to catch this entity
-        .with(FlyControlTag)
-        // Will allow the ArcBallMovementSystem to catch this entity
-        .with(ArcBallControlTag {
-            target: entity,
-            distance: 10.,
-        })
-        .build();
 }