+++ src/app.rs
@@ -389,7 +389,9 @@ impl<S> ApplicationBuilder<S> {
             .level(LevelFilter::Debug)
             .chain(io::stdout())
             .apply()
-            .unwrap_or_else(|_| debug!("Global logger already set, default amethyst logger will not be used"));
+            .unwrap_or_else(|_| {
+                debug!("Global logger already set, default amethyst logger will not be used")
+            });
 
         info!("Initializing Amethyst...");
         info!("Version: {}", env!("CARGO_PKG_VERSION"));